requirements.txt
folium
PyYAML
gpxpy
geopy